legacy_backend/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
Flask-Login==0.6.3
Werkzeug==3.0.1
gunicorn==21.2.0
supabase==2.4.4
httpx>=0.26.0
pytz==2023.3.post1
